% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopingreview-package.R
\docType{package}
\name{scopingreview-package}
\alias{scopingreview}
\alias{scopingreview-package}
\title{scopingreview: Automated Systematic and Scoping Literature Reviews via Scopus and OpenAlex}
\description{
Builds Boolean search queries from topic keywords and optional geographic filters, retrieves records from the Elsevier Scopus and OpenAlex APIs, harmonises their metadata, removes duplicates using exact DOI and fuzzy title matching, and flags papers published in a curated list of political science, public administration and African studies journals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PPADEM/scopingreview}
  \item Report bugs at \url{https://github.com/PPADEM/scopingreview/issues}
}

}
\author{
\strong{Maintainer}: Josh Tyler \email{josh.tyler@bristol.ac.uk}

Authors:
\itemize{
  \item Josh Tyler \email{josh.tyler@bristol.ac.uk}
}

}
\keyword{internal}
