% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{TARGET_JOURNALS}
\alias{TARGET_JOURNALS}
\title{Curated target journal list}
\format{
A character vector of journal names.
}
\usage{
TARGET_JOURNALS
}
\description{
Political science, public administration and area/African studies journals
used by \code{\link[=deduplicate_records]{deduplicate_records()}} to set the \code{target_journal} flag. Pass your
own character vector to \code{target_journals} to use a different list.
}
